+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+Breaking Changes:
+
+- [task] Widened the scope of some methods in TaskManager and TaskConfigurations from string to TaskConfigurationScope. This is only breaking for extenders, not callers. [#7928](https://github.com/eclipse-theia/theia/pull/7928)
+
 ## v1.2.0
 
 - [application-manager] added ability for clients to add `windowOptions` using an IPC-Event [#7803](https://github.com/eclipse-theia/theia/pull/7803)